+++ CHANGES.rst
@@ -22,6 +22,9 @@ Changelog
   - Reuse the Python mulitprocessing pool to avoid the overhead of spawning
     new workers for each parallel section.
 
+  - Add chunk size heuristic to reduce overhead of parallel maps using Python
+    mulitprocessing.
+
 2.1.2 (2024-10-22)
 ==================